Added: distance_get.html, distance_post.html and DistanceSearchController Updated:  build.gradle with org.json implementation
Added a search page "/Distance" that lets user enter an address and see the time and distance to current provider.

Still needed:  Validation to ensure addresses are in the proper format.

Utilizes the Distance Matrix API

Since this API is not free after the initial 800 elements a direct link to this function won't be made available on the home page.  This is for proof of concept.